+++ packages/use-shortcut/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Removed non-React public entry points (`createShortcut`, `createShortcutMap`) to focus package API on React-first usage.
+
 ## [1.3.0] - 2026-02-28
 
 ### Added